services/discord_service/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "discord_service"
version = "0.1.0"
description = "Bibliotalk Discord runtime"
requires-python = ">=3.11"
dependencies = [
  "agents_service>=0.1.0",
  "bt_common>=0.1.0",
  "bt_store>=0.1.0",
  "discord.py>=2.4.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.23.8",
  "ruff>=0.6.9",
]

[tool.setuptools]
packages = [
  "discord_service",
  "discord_service.bot",
  "discord_service.feed",
  "discord_service.ops",
  "discord_service.talks",
]

[tool.setuptools.package-dir]
"discord_service" = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.uv.sources]
agents-service = { workspace = true }
bt-common = { workspace = true }
bt-store = { workspace = true }
